Extract `useMagicCards` to composable
